add json web token seems good, next need to process the img on server

+++ js/router/main.js
@@ -11,6 +11,7 @@ var urlencodedParser = bodyParser.urlencoded({ extended: false })
 var qiniu = require("qiniu");
 const multiparty = require('multiparty');
 const fs = require('fs');
+const JwtUtil = require('../jwt');
 let router = express.Router();
 var app = express();
 
@@ -47,25 +48,36 @@ let query = function( sql, values ) {
   app.use(bodyParser.json());//数据JSON类型
   
  
-
+function getid(token) {
+   
+   var jwt = new JwtUtil(token);
+   var ret;
+   try{
+      ret = jwt.verifyToken().id;
+   }catch(e){
+       ret = "-1";
+   }
+   return(ret)
+}
 
 
 router.post('/getdata',urlencodedParser, async (req, res) => {
-    var username = req.body.username;
-    console.log("username:"+username+" is ask for data");
+    
+    userid = getid(req.body.token)
+    if(userid=="-1"){
+       res.status(400).send("token expired")
+       return;
+    }
+    console.log("userid:"+userid+" is ask for data");
  
-    const result = await query("select lastopenfileid from user where username = '"+username+"'");
+    const result = await query("select lastopenfileid from user where userid = '"+userid+"'");
     
     if(result.length==1){
- 
-       const result3 = await query("select userid from user where username = '"+username+"'");
-    
- 
-       
+   
        // console.log(result) 这里即使result是null也会有数据返回 解决方法是 用户注册完 在note里生成一个startnote.md 作为用户的lastopenfileid
-       const result2 = await query("select content from note where fileid = '"+result[0].lastopenfileid+"'");
+       const result2 = await query("select content from note where fileid = '"+result[0].lastopenfileid+"' and userid="+userid);
        if(result2.length>=1){
-          var x = {content:result2[0].content,id:result[0].lastopenfileid,userid:result3[0].userid}
+          var x = {content:result2[0].content,id:result[0].lastopenfileid,userid:userid}
           res.send(x);
        }
     }
@@ -75,33 +87,33 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
  
  
  router.post('/getfile',urlencodedParser, function (req, res) {
-    var username = req.body.username;
     var fileid = req.body.fileid;
-    console.log("username:"+username+" is ask for file" + fileid);
+    userid = getid(req.body.token)
+    if(userid=="-1"){
+       res.status(400).send("token expired")
+       return;
+    }
+    console.log("userid:"+userid+" is ask for file");
  
     pool.getConnection(function(err,connection){
-       console.log("getfile : connection to sql success")
-       var qu = "select userid from user where username = '"+username+"'";
-       connection.query(qu,function(err,result){
-          if(result.length==1){
-             qu = "select content from note where userid = '"+result[0].userid+"' and fileid ='" +fileid +"'";
+      //  console.log("getfile : connection to sql success")
+       
+          
+             qu = "select content from note where userid = '"+userid+"' and fileid ='" +fileid +"'";
              connection.query(qu,function(err,result2){
-                if(result.length>=1){
-                   var x = {content:result2[0].content,fileid:fileid,userid:result[0].userid}
+                if(result2.length>=1){
+                   var x = {content:result2[0].content,fileid:fileid,userid:userid}
                    res.send(x);
                 }
                 else{
-                   console.log("getfile : connection to mysql fail")
+                  //  console.log("getfile : connection to mysql fail")
                 }    
                 })
              connection.release();
-             console.log("getfile : mysql release")
-          }
-          else{
-             console.log("getfile : connection to mysql fail")
-             connection.release();
-          }    
-          })
+            //  console.log("getfile : mysql release")
+          
+             
+         
        })
  
  })
@@ -111,34 +123,36 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
  
  
  router.post('/getcatalogue',urlencodedParser,  async (req, res) => {
-    var username = req.body.username;
-    console.log("username:"+username+" is ask for content");
+   userid = getid(req.body.token)
+   if(userid=="-1"){
+      res.status(400).send("token expired")
+      return;
+   }
+   console.log("userid:"+userid+" is ask for catalogue");
  
  
-    const result = await query("select userid from user where username = '"+username+"'");
- 
-    if(result.length>=1){
-       const result2 = await query("select filename,isnote,level,fileid,fatherid from catalogue where userid = "+result[0].userid);
-       if(result2.length>=1){
-          res.send(result2)
-       }  
-       else{
-          console.log(result2)
-       }
-    }   
-    else{
-       console.log(result)
-    }
+   const result2 = await query("select filename,isnote,level,fileid,fatherid from catalogue where userid = "+userid);
+   if(result2.length>=1){
+      res.send(result2)
+   }  
+   else{
+      console.log(result2)
+   }
+    
  })
  
  router.post('/savedata',urlencodedParser, function (req, res) {
     var content = req.body.content;
     var id = req.body.id;
-    var userid = req.body.userid;
-    console.log("is saving data for file " + id);
+    userid = getid(req.body.token)
+    if(userid=="-1"){
+       res.status(400).send("token expired")
+       return;
+    }
+    console.log("userid:"+userid+" is saving file");
  
     pool.getConnection(function(err,connection){
-       console.log("savedata : connection to sql success")
+      //  console.log("savedata : connection to sql success")
        var qu = "update note set content ='" + content + "'where fileid ="+id +" and userid="+userid+";";
        connection.query(qu,function(err,result){
           if(err){
@@ -152,10 +166,10 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
                 return;
              } 
              connection.release();
-             console.log("savedata : mysql release")
+            //  console.log("savedata : mysql release")
              res.send("success");
              })
-          console.log("savedata : mysql release")
+         //  console.log("savedata : mysql release")
           })
        })
       
@@ -167,11 +181,15 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
  // deletefile
  router.post('/deletefile',urlencodedParser, function (req, res) {
     var fileid = req.body.fileid;
-    var userid = req.body.userid;
-    console.log("deletefolder id:" + fileid);
+    userid = getid(req.body.token)
+    if(userid=="-1"){
+       res.status(400).send("token expired")
+       return;
+    }
+    console.log("userid:"+userid+" is deleting folder id:" + fileid);
  
     pool.getConnection(function(err,connection){
-       console.log("deletefolder : connection to sql success")
+      //  console.log("deletefolder : connection to sql success")
        var qu = "delete from catalogue where fileid = " + fileid + " and userid="+userid+";";
        connection.query(qu,function(err,result){
           if(err){
@@ -186,16 +204,15 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
                 return;
              } 
              
-             console.log("delete a file in catalogue")
-          })
+          });
           var qu3 = "delete from note where fatherid = " +fileid+ " and userid="+userid+";";
           connection.query(qu3,function(err3,result3){
              if(err3){
                 console.log('[UPDATE ERROR] - ',err3.message);
                 return;
              } 
-          })
-          })
+          });
+          });
           var qu4 = "delete from note where fileid = " +fileid+ " and userid="+userid+";";
           connection.query(qu4,function(err4,result4){
              if(err4){
@@ -203,12 +220,12 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
                 return;
              } 
              res.send("success");
-          })
+          });
           connection.release();
-     })
+     });
        
  
- })
+ });
  
  
  
@@ -224,11 +241,15 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
  
  // newfolder
  router.post('/newfolder',urlencodedParser, function (req, res) {
-    var userid = req.body.userid;
-    console.log("new a folder for  user:" + userid);
- 
+   userid = getid(req.body.token)
+   if(userid=="-1"){
+      res.status(400).send("token expired")
+      return;
+   }
+   console.log("userid:"+userid+" is newing folder");
+
     pool.getConnection(function(err,connection){
-       console.log("newfolder : connection to sql success")
+      //  console.log("newfolder : connection to sql success")
        var qu = "select max(fileid) from catalogue where userid="+userid;
        connection.query(qu,function(err,result){
           if(err){
@@ -244,12 +265,12 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
              } 
              
              res.json(((result[0]["max(fileid)"]+1)));
-             })
+             });
           connection.release();
-          })
-       })
+          });
+       });
  
- })
+ });
  
  
  
@@ -258,12 +279,17 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
  
  // newfile
  router.post('/newfile',urlencodedParser, function (req, res) {
-    var userid = req.body.userid;
+   userid = getid(req.body.token)
+   if(userid=="-1"){
+      res.status(400).send("token expired")
+      return;
+   }
+   console.log("userid:"+userid+" is newing file");
+
     var folderid = req.body.folderid;
-    console.log("new a  file for  user:" + userid);
  
     pool.getConnection(function(err,connection){
-       console.log("newfolder : connection to sql success")
+      //  console.log("newfolder : connection to sql success");
        var qu = "select max(fileid) from catalogue where userid="+userid;
        connection.query(qu,function(err,result){
           if(err){
@@ -277,7 +303,7 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
                 console.log('[UPDATE ERROR] - ',err2.message);
                 return;
              } 
-             })
+             });
           var qu3 = "insert into note values("+(result[0]["max(fileid)"]+1)+",'# newfile',"+userid+","+folderid +");";
           connection.query(qu3,function(err3,result3){
              if(err3){
@@ -286,40 +312,42 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
              } 
              
              
-             })
+             });
           connection.release();
  
           res.json(((result[0]["max(fileid)"]+1)));
-          })
-       })
+          });
+       });
  
- })
- 
- 
+ });
  
  
  
  
  
- router.post('/savecatalogue',urlencodedParser, function (req, res) {
+ 
+ 
+ router.post('/savecatalogue',urlencodedParser, function (req, res) {//rename
     var filename = req.body.filename;
     var fileid = req.body.fileid;
-    console.log("is saving catalogue for file " + fileid);
+    userid = getid(req.body.token)
+   if(userid=="-1"){
+      res.status(400).send("token expired")
+      return;
+   }
+   console.log("userid:"+userid+" is saving catalogue for file " + fileid);
  
     pool.getConnection(function(err,connection){
-       console.log("savecatalogue : connection to sql success")
-       var qu = "update catalogue set filename ='" + filename + "'where fileid ="+fileid;
+      //  console.log("savecatalogue : connection to sql success")
+       var qu = "update catalogue set filename ='" + filename + "'where fileid ="+fileid+" and userid ="+userid;
        connection.query(qu,function(err,result){
           if(err){
              console.log('[UPDATE ERROR] - ',err.message);
              return;
           } 
-          console.log("savecatalogue : mysql release")
+         //  console.log("savecatalogue : mysql release")
           res.send("success");
  
- 
- 
-          
           connection.release();
           })
        })
@@ -338,26 +366,35 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
  
  router.post("/upload",urlencodedParser,  async (req, res) => {
    let form = new multiparty.Form();
+   
    form.parse(req, async(err,fields,file)=>{
-      const result = await query("select max(fileid) from catalogue where userid="+fields["userid"])
+      userid = getid(fields["token"][0])
+      if(userid=="-1"){
+         res.status(400).send("token expired")
+         console.log("token expired!!!!!")
+         return;
+      }
+      console.log("userid:"+userid+" is uploading file(md)");
+      const result = await query("select max(fileid) from catalogue where userid="+userid)
       var maxindex = (result[0]["max(fileid)"])
-      console.log("maxindex now  ：  "+maxindex)
+      console.log("maxindex in catalogue ：  "+maxindex)
       if(fields["folderid"]==-1){//不是目录，先创建一个目录
-         result2 = await query( "INSERT INTO catalogue VALUES("+fields["userid"] +","+(maxindex+1)+",'newfolder',0,0,null)")
+         result2 = await query( "INSERT INTO catalogue VALUES("+userid +","+(maxindex+1)+",'newfolder',0,0,null)")
          for (let index = 2; index < (parseInt(fields["size"])+2); index++) {//每个文件存入该目录  
-            orifilename = file["file"][index-1]["originalFilename"];
+            orifilename = file["file"][index-2]["originalFilename"];
             filename = orifilename.substring(0,orifilename.indexOf("."))
             endname = orifilename.substring(orifilename.indexOf("."),orifilename.length)
             if(endname != ".md") 
                return;
-            result2 = await query( "INSERT INTO catalogue VALUES("+fields["userid"] +","+(maxindex+index)+",'"+filename+"',1,1,"+(maxindex+1)+")")
+            result2 = await query( "INSERT INTO catalogue VALUES("+userid +","+(maxindex+index)+",'"+filename+"',1,1,"+(maxindex+1)+")")
             var data = fs.readFileSync(file["file"][index-2]["path"], 'utf-8');
             data = data.replace(/\\/g,"\\\\");
                 data = data.replace(/\"/g,"\'\'");
                 data =data.replace(/\'/g,"\"");
-            result3 = await query( "INSERT INTO note VALUES("+(maxindex+index)+",'"+data+"',"+fields["userid"]+","+(maxindex+1)+")")
-            res.send("success")
+            result3 = await query( "INSERT INTO note VALUES("+(maxindex+index)+",'"+data+"',"+userid+","+(maxindex+1)+")")
+            
          }
+         res.send("success")
       }else{
          for (let index = 1; index < (parseInt(fields["size"])+1); index++) {//每个文件存入该目录  
             orifilename = file["file"][index-1]["originalFilename"];
@@ -366,12 +403,12 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
             if(endname != ".md") 
                return;
             fatherid = fields["folderid"]
-            result2 = await query( "INSERT INTO catalogue VALUES("+fields["userid"] +","+(maxindex+index)+",'"+filename+"',1,1,"+(fatherid)+")")
+            result2 = await query( "INSERT INTO catalogue VALUES("+userid +","+(maxindex+index)+",'"+filename+"',1,1,"+(fatherid)+")")
             var data = fs.readFileSync(file["file"][index-1]["path"], 'utf-8');
             data = data.replace(/\\/g,"\\\\");
                 data = data.replace(/\"/g,"\'\'");
                 data =data.replace(/\'/g,"\"");
-            result3 = await query( "INSERT INTO note VALUES("+(maxindex+index)+",'"+data+"',"+fields["userid"]+","+(fatherid)+")")
+            result3 = await query( "INSERT INTO note VALUES("+(maxindex+index)+",'"+data+"',"+userid+","+(fatherid)+")")
             res.send("success")
          }
       }
@@ -379,7 +416,13 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
  });
 
  router.post("/uploadimg",urlencodedParser,  async (req, res) => {
-      var userid = req.body.userid;
+   userid = getid(req.body.token)
+   if(userid=="-1"){
+      res.status(400).send("token expired")
+      return;
+   }
+   console.log("userid:"+userid+" is uploading img");
+ 
       var fileid = req.body.fileid;
       var size = req.body.size;
       var hash = req.body.hash;
@@ -388,8 +431,16 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
       const result3 = await query("insert into img values("+fileid+","+userid+",'"+hash+"',"+size+") ")
  });
 
+
+ 
  router.post("/getleftsize",urlencodedParser,  async (req, res) => {
-   var userid = req.body.userid;
+   userid = getid(req.body.token)
+   if(userid=="-1"){
+      res.status(400).send("token expired")
+      return;
+   }
+   console.log("userid:"+userid+" is asking for left size");
+ 
    const result = await query("select maxsize from user where userid="+userid)
    const result1 = await query("select size from user where userid="+userid)
    ret = parseInt(result[0].maxsize) - parseInt(result1[0].size);
@@ -403,7 +454,6 @@ router.post('/getdata',urlencodedParser, async (req, res) => {
  
  
 router.get('/',urlencodedParser, function (req, res) {
-   console.log("!")
    res.status(200).sendFile( path.resolve(__dirname + "/../../public/html/" + "notepage.html") );
 });
